requirements.txt
fastapi
uvicorn[standard]
pywebview
requests
